+++ presidio-image-redactor/presidio_image_redactor/dicom_image_redactor_engine.py
@@ -26,9 +26,10 @@ class DicomImageRedactorEngine(ImageRedactorEngine):
     def redact(
         self,
         image: pydicom.dataset.FileDataset,
-        fill: Optional[str] = "contrast",
-        padding_width: Optional[int] = 25,
-        crop_ratio: Optional[float] = 0.75,
+        fill: str = "contrast",
+        padding_width: int = 25,
+        crop_ratio: float = 0.75,
+        ocr_kwargs: Optional[dict] = None,
         **kwargs,
     ):
         """Redact method to redact the given DICOM image.
@@ -41,6 +42,7 @@ class DicomImageRedactorEngine(ImageRedactorEngine):
         :param padding_width: Padding width to use when running OCR.
         :param crop_ratio: Portion of image to consider when selecting
         most common pixel value as the background color value.
+        :param ocr_kwargs: Additional params for OCR methods.
         :param kwargs: Additional values for the analyze method in AnalyzerEngine
 
         :return: DICOM instance with redacted pixel data.
@@ -67,6 +69,7 @@ class DicomImageRedactorEngine(ImageRedactorEngine):
         analyzer_results = self.image_analyzer_engine.analyze(
             image,
             ad_hoc_recognizers=[deny_list_recognizer],
+            ocr_kwargs=ocr_kwargs,
             **kwargs,
         )
 
@@ -80,9 +83,10 @@ class DicomImageRedactorEngine(ImageRedactorEngine):
         self,
         input_dicom_path: str,
         output_dir: str,
-        padding_width: Optional[int] = 25,
-        crop_ratio: Optional[float] = 0.75,
-        fill: Optional[str] = "contrast",
+        padding_width: int = 25,
+        crop_ratio: float = 0.75,
+        fill: str = "contrast",
+        ocr_kwargs: Optional[dict] = None,
         **kwargs,
     ) -> None:
         """Redact method to redact from a given file.
@@ -95,6 +99,7 @@ class DicomImageRedactorEngine(ImageRedactorEngine):
         :param padding_width : Padding width to use when running OCR.
         :param fill: Color setting to use for redaction box
         ("contrast" or "background").
+        :param ocr_kwargs: Additional params for OCR methods.
         :param kwargs: Additional values for the analyze method in AnalyzerEngine
         """
         # Verify the given paths
@@ -118,6 +123,7 @@ class DicomImageRedactorEngine(ImageRedactorEngine):
             padding_width=padding_width,
             overwrite=True,
             dst_parent_dir=".",
+            ocr_kwargs=ocr_kwargs,
             **kwargs,
         )
 
@@ -129,9 +135,10 @@ class DicomImageRedactorEngine(ImageRedactorEngine):
         self,
         input_dicom_path: str,
         output_dir: str,
-        padding_width: Optional[int] = 25,
-        crop_ratio: Optional[float] = 0.75,
-        fill: Optional[str] = "contrast",
+        padding_width: int = 25,
+        crop_ratio: float = 0.75,
+        fill: str = "contrast",
+        ocr_kwargs: Optional[dict] = None,
         **kwargs,
     ) -> None:
         """Redact method to redact from a directory of files.
@@ -146,6 +153,7 @@ class DicomImageRedactorEngine(ImageRedactorEngine):
         most common pixel value as the background color value.
         :param fill: Color setting to use for redaction box
         ("contrast" or "background").
+        :param ocr_kwargs: Additional params for OCR methods.
         :param kwargs: Additional values for the analyze method in AnalyzerEngine
         """
         # Verify the given paths
@@ -169,6 +177,7 @@ class DicomImageRedactorEngine(ImageRedactorEngine):
             padding_width=padding_width,
             overwrite=True,
             dst_parent_dir=".",
+            ocr_kwargs=ocr_kwargs,
             **kwargs,
         )
 
@@ -760,6 +769,7 @@ class DicomImageRedactorEngine(ImageRedactorEngine):
         padding_width: int,
         overwrite: bool,
         dst_parent_dir: str,
+        ocr_kwargs: Optional[dict] = None,
         **kwargs,
     ) -> str:
         """Redact text PHI present on a DICOM image.
@@ -773,6 +783,7 @@ class DicomImageRedactorEngine(ImageRedactorEngine):
         :param overwrite: Only set to True if you are providing the
         duplicated DICOM path in dcm_path.
         :param dst_parent_dir: String path to parent directory of where to store copies.
+        :param ocr_kwargs: Additional params for OCR methods.
         :param kwargs: Additional values for the analyze method in AnalyzerEngine
 
         :return: Path to the output DICOM file.
@@ -809,6 +820,7 @@ class DicomImageRedactorEngine(ImageRedactorEngine):
         analyzer_results = self.image_analyzer_engine.analyze(
             image,
             ad_hoc_recognizers=[deny_list_recognizer],
+            ocr_kwargs=ocr_kwargs,
             **kwargs,
         )
 
@@ -829,6 +841,7 @@ class DicomImageRedactorEngine(ImageRedactorEngine):
         padding_width: int,
         overwrite: bool,
         dst_parent_dir: str,
+        ocr_kwargs: Optional[dict] = None,
         **kwargs,
     ) -> str:
         """Redact text PHI present on all DICOM images in a directory.
@@ -842,6 +855,7 @@ class DicomImageRedactorEngine(ImageRedactorEngine):
         :param overwrite: Only set to True if you are providing
         the duplicated DICOM dir in dcm_dir.
         :param dst_parent_dir: String path to parent directory of where to store copies.
+        :param ocr_kwargs: Additional params for OCR methods.
         :param kwargs: Additional values for the analyze method in AnalyzerEngine
 
         Return:
@@ -869,6 +883,7 @@ class DicomImageRedactorEngine(ImageRedactorEngine):
                 padding_width,
                 overwrite,
                 dst_parent_dir,
+                ocr_kwargs=ocr_kwargs,
                 **kwargs,
             )