+++ presidio-image-redactor/presidio_image_redactor/image_analyzer_engine.py
@@ -1,4 +1,4 @@
-from typing import List, Optional
+from typing import List, Tuple, Optional
 
 from presidio_analyzer import AnalyzerEngine, RecognizerResult
 
@@ -27,23 +27,26 @@ class ImageAnalyzerEngine:
             ocr = TesseractOCR()
         self.ocr = ocr
 
-    def analyze(self, image: Optional[object], **kwargs) -> List[ImageRecognizerResult]:
+    def analyze(
+        self, image: Optional[object], ocr_kwargs: Optional[dict] = None, **kwargs
+    ) -> List[ImageRecognizerResult]:
         """Analyse method to analyse the given image.
 
-        :param image: PIL Image/numpy array or file path(str) to be processed
-        :param kwargs: Additional values for the analyze method in AnalyzerEngine
+        :param image: PIL Image/numpy array or file path(str) to be processed.
+        :param ocr_kwargs: Additional params for OCR methods.
+        :param kwargs: Additional values for the analyze method in AnalyzerEngine.
 
-        :return: list of the extract entities with image bounding boxes
+        :return: List of the extract entities with image bounding boxes.
         """
-        ocr_result = self.ocr.perform_ocr(image)
+        # Perform OCR
+        perform_ocr_kwargs, ocr_threshold = self._parse_ocr_kwargs(ocr_kwargs)
+        ocr_result = self.ocr.perform_ocr(image, **perform_ocr_kwargs)
 
         # Apply OCR confidence threshold if it is passed in
-        if "ocr_threshold" in kwargs:
-            ocr_result = self.threshold_ocr_result(ocr_result, kwargs["ocr_threshold"])
-            kwargs = {
-                key: value for key, value in kwargs.items() if key != "ocr_threshold"
-            }
+        if ocr_threshold is not None:
+            ocr_result = self.threshold_ocr_result(ocr_result, ocr_threshold)
 
+        # Analyze text
         text = self.ocr.get_text_from_ocr_dict(ocr_result)
 
         analyzer_result = self.analyzer_engine.analyze(
@@ -58,12 +61,10 @@ class ImageAnalyzerEngine:
     def threshold_ocr_result(ocr_result: dict, ocr_threshold: float) -> dict:
         """Filter out OCR results below confidence threshold.
 
-        Args:
-            ocr_result (dict): OCR results (raw).
-            ocr_threshold (float): Threshold value between -1 and 100.
+        :param ocr_result: OCR results (raw).
+        :param ocr_threshold: Threshold value between -1 and 100.
 
-        Return:
-            filtered_ocr_result (dict): OCR results with low confidence items removed.
+        :return: OCR results with low confidence items removed.
         """
         if ocr_threshold < -1 or ocr_threshold > 100:
             raise ValueError("ocr_threshold must be between -1 and 100")
@@ -155,3 +156,25 @@ class ImageAnalyzerEngine:
                 pos += len(word) + 1
 
         return bboxes
+
+    @staticmethod
+    def _parse_ocr_kwargs(ocr_kwargs: dict) -> Tuple[dict, float]:
+        """Parse the OCR-related kwargs.
+
+        :param ocr_kwargs: Parameters for OCR operations.
+
+        :return: Params for ocr.perform_ocr and ocr_threshold
+        """
+        ocr_threshold = None
+        if ocr_kwargs is not None:
+            if "ocr_threshold" in ocr_kwargs:
+                ocr_threshold = ocr_kwargs["ocr_threshold"]
+                ocr_kwargs = {
+                    key: value
+                    for key, value in ocr_kwargs.items()
+                    if key != "ocr_threshold"
+                }
+        else:
+            ocr_kwargs = {}
+
+        return ocr_kwargs, ocr_threshold