+++ presidio-image-redactor/presidio_image_redactor/image_pii_verify_engine.py
@@ -24,20 +24,23 @@ class ImagePiiVerifyEngine:
             image_analyzer_engine = ImageAnalyzerEngine()
         self.image_analyzer_engine = image_analyzer_engine
 
-    def verify(self, image: Image, **kwargs) -> Image:
+    def verify(
+        self, image: Image, ocr_kwargs: Optional[dict] = None, **kwargs
+    ) -> Image:
         """Annotate image with the detect PII entity.
 
         Please notice, this method duplicates the image, creates a
         new instance and manipulate it.
 
-        :param image: PIL Image to be processed
+        :param image: PIL Image to be processed.
+        :param ocr_kwargs: Additional params for OCR methods.
         :param kwargs: Additional values for the analyze method in ImageAnalyzerEngine.
         :return: the annotated image
         """
 
         image = ImageChops.duplicate(image)
         image_x, image_y = image.size
-        bboxes = self.image_analyzer_engine.analyze(image, **kwargs)
+        bboxes = self.image_analyzer_engine.analyze(image, ocr_kwargs, **kwargs)
         fig, ax = plt.subplots()
         image_r = 70
         fig.set_size_inches(image_x / image_r, image_y / image_r)