+++ presidio-image-redactor/presidio_image_redactor/image_redactor_engine.py
@@ -1,4 +1,4 @@
-from typing import Union, Tuple
+from typing import Union, Tuple, Optional
 
 from PIL import Image, ImageDraw, ImageChops
 
@@ -18,8 +18,10 @@ class ImageRedactorEngine:
             self.image_analyzer_engine = image_analyzer_engine
 
     def redact(
-        self, image: Image,
+        self,
+        image: Image,
         fill: Union[int, Tuple[int, int, int]] = (0, 0, 0),
+        ocr_kwargs: Optional[dict] = None,
         **kwargs,
     ) -> Image:
         """Redact method to redact the given image.
@@ -29,6 +31,7 @@ class ImageRedactorEngine:
         :param image: PIL Image to be processed
         :param fill: colour to fill the shape - int (0-255) for
         grayscale or Tuple(R, G, B) for RGB
+        :param ocr_kwargs: Additional params for OCR methods.
         :param kwargs: Additional values for the analyze method in AnalyzerEngine
 
         :return: the redacted image
@@ -36,7 +39,7 @@ class ImageRedactorEngine:
 
         image = ImageChops.duplicate(image)
 
-        bboxes = self.image_analyzer_engine.analyze(image, **kwargs)
+        bboxes = self.image_analyzer_engine.analyze(image, ocr_kwargs, **kwargs)
         draw = ImageDraw.Draw(image)
 
         for box in bboxes: