+++ CHANGELOG.md
@@ -7,7 +7,7 @@ All notable changes to this project will be documented in this file.
 #### Image redactor
 * Added evaluation code for the DICOM image redaction capabilities
 * Modified `ImagePiiVerifyEngine` to allow passing of kwargs
-* Updated `ImageAnalyzerEngine` to allow passing of an OCR confidence threshold
+* Updated all image redactor engines and OCR classes to allow passing of an OCR confidence threshold and other OCR parameters.
 
 #### General
 * Updated documentation to include instructions on using DICOM evaluation code